+++ test/runTests.ts
@@ -2,5 +2,6 @@ import './cellView.test';
 import './clipboard.test';
 import './dataReader.test';
 import './messages.test';
+import './profile.test';
 import './tableLayout.test';
 import './tableState.test';